Data.SignTyped.Class

High-level entrypoints

hashSignable :: forall a baseType. SignableValueDefault baseType a => a -> Either String ByteString

Perform EIP-712 encodeData on the given value with its type's default handle.

hashSignableWith :: forall a. SignableValueHandle a -> a -> Either String ByteString

Perform EIP-712 encodeData on the given value using the given handle.

Typeclasses

class SignableValue (baseType :: Symbol) a where

SignableValue baseType a allows a Haskell value of type a to be represented as a Solidity value with the typename baseType.

The relation between Solidity and Haskell types is many-to-many. For example, Solidity's uint64 can be produced from any of Word, Int, Natural or Integer. On the other hand, Haskell's Integer can be mapped to any of Solidity's numeric types.

The actual implementation is delegated to a SignableValueHandle object, which allows us some additional flexibility:

The SignableValue itself is treated as a default implementation rather than the canonical one.

To define an instance of SignableValue, one only needs to perform a single assignment, most often from a handle-template function, instead of repeating the same parameters across multiple class method definitions.

Minimal complete definition

signableValueHandle

Methods

signableValueHandle :: SignableValueHandle a

signableValueListHandle :: SignableValueHandle [a]

class SignableValue baseType a => SignableValueDefault baseType a | a -> baseType

This typeclass sets one of the Solidity typenames as the default correspondence for the given Haskell type.

signableValueDefaultHandle :: forall a baseType. SignableValueDefault baseType a => SignableValueHandle a

A version of signableValueHandle that uses the stronger SignableValueDefault constraint.

Generic implementation

newtype GenericSignable (mods :: [SigningModifier]) a

This newtype is intended to be used as a wrapper for DerivingVia to derive SignableValue, for example:

       data MyPerson = MyPerson
           { pname :: String
           , pwallet :: BS.ByteString
           }
           deriving (Show, Eq, Generic)
           deriving (SignableValue "Person") via (
               GenericSignable
                   '[  'SigningFieldModifier "pname"
                           ('Just "name")
                           ('Just (AsBaseType "string" String))
                   ,   'SigningFieldModifier "pwallet"
                           ('Just "wallet")
                           ('Just (AsBaseType "address" BS.ByteString))
                   ]
                   MyPerson
           )
           deriving anytype (SignableValueDefault "Person")
   

mods is a DataKinds parameter that is used to fine-tune the generated instance, similar to aeson's Options, but on a type level instead of the value level.

Constructors

GenericSignable

Fields

unwrapGenericSignable :: AssertValidation (ValidateModifiers mods (Rep a)) a

data SigningModifier

A datatype for GenericSignable modifiers. It's intended to be used at type level with DataKinds.

Each modifier in the list can be thought of as an assignment of the provided value to a configuration parameter.

SigningFieldModifier "haskellFieldName" mbEthName mbSignType configures the field, which is referred to in Haskell by the selector haskellFieldName. mbEthName and mbSignType are optional new configuration values: 'Just smth sets the configuration to the specified value, while 'Nothing leaves it unchanged.

The second parameter, mbEthName, sets the name of the field in Solidity. By default, the Haskell selector's name is used.

The third paremeter, mbSignType, is a DerivingVia-style wrapper that configures a "signing type". This type must have a valid SignableValueDefault instance, and, obviously, be Coercible with the actual field's type.

Constructors

SigningFieldModifier Symbol (Maybe Symbol) (Maybe Type)

genericSignableValueHandle :: forall baseType mods s. (Generic s, AssertValidation (ValidateModifiers mods (Rep s)) (() :: Constraint), GSignableRep mods s (Rep s), KnownSymbol baseType) => SignableValueHandle s

A generic implementation of the signable handle. You can use this function instead of GenericSignable to define your own handle or instance, for example:

       data MyPerson = MyPerson
           { pname :: String
           , pwallet :: BS.ByteString
           }
           deriving (Show, Eq, Generic)

       myPersonSignableValueHandle :: SignableValueHandle MyPerson
       myPersonSignableValueHandle =
           genericSignableValueHandle
               @"Person"
               @'[ 'SigningFieldModifier "pname"
                       ('Just "name")
                       ('Just (AsBaseType "string" String))
               ,   'SigningFieldModifier "pwallet"
                       ('Just "wallet")
                       ('Just (AsBaseType "address" BS.ByteString))
               ]

       instance SignableValue "Person" MyPerson where
           signableValueHandle = myPersonSignableValueHandle

       instance SignableValueDefault "Person" MyPerson where {}
   

Modifier wrappers

newtype AsBaseType (baseType :: Symbol) a

This is a newtype wrapper that's indended to be used with SigningFieldModifier.

It sets the Solidity typename to be used for this Haskell type. This allows you to use a SignableValue instance that is different from default, or on types that don't even define a default, such as Integer and ByteString.

Constructors

AsBaseType a

newtype AsSized (arrSize :: Nat) a

This is a newtype wrapper that's indended to be used with SigningFieldModifier.

Being a wrapper for an array, it allows you to set a fixed size to the array, as opposed to having the default option to map the Haskell list to a variable-length array.

For example, representing a Solidity structure Rectangle {int256[2] minpos; int256[2] maxpos} could be done like this:

       data Rectangle = Rectangle
           { minpos :: [Integer]
           , maxpos :: [Integer]
           }
           deriving (Generic)
           deriving (SignableValue "Rectangle") via (
               GenericSignable
                   '[  'SigningFieldModifier "minpos"
                           'Nothing
                           ('Just (AsSized 2 (AsBaseType "int256" Integer)))
                   ,   'SigningFieldModifier "maxpos"
                           'Nothing
                           ('Just (AsSized 2 (AsBaseType "int256" Integer)))
                   ]
                   Rectangle
           )
   

Constructors

AsSized [a]

Simple value handles

boolValueHandle :: SignableValueHandle Bool

This handle encodes Haskell Bool into Solidity's bool:

False is encoded as replicate 32 0.

True is encoded as replicate 31 0 <> 'x01'.

numberValueHandle :: AtomicIntegerSignedness -> AtomicIntegerBitSize -> SignableValueHandle Integer

This handle encodes Haskell Integer into a Solidity's integer type: sign-extended to 256 bits (32 bytes) and serialized in the big-endian order.

The two parameters define signedness and bit size of the integer.

bytesValueHandle

Arguments

:: AtomicBytesSubtype
The subtype of the bytestring.

-> SignableValueHandle ByteString

This handle encodes Haskell ByteString into Solidity's address, bytesN or bytes.

The single parameter chooses a subtype:

AtomicBytesSubtypeAddress allows bytestrings of length 20, and encodes it as an address: left-padded with zero bytes until the length of 32.

AtomicBytesSubtypeShortBytes N allows bytestrings with length of exactly N, and encodes it as a bytesN: right-padded with zero bytes until the length of 32. The handle will only make a valid encoding for 1 <= N <= 32.

AtomicBytesSubtypeLongBytes allows bytestrings of any length and any contents, and encodes it as bytes: by taking the keccak256 of the bytes.

stringValueHandle :: SignableValueHandle Text

This handle encodes Haskell Text into Solidity's string: by taking the keccak256 of the UTF-8 encoding of the string.

arrayValueHandle

Arguments

:: SignableValueHandle a
Underlying type's handle.

-> Maybe Int
Array size.

-> SignableValueHandle [a]

This handle encodes the array type.

The first parameter is the handle for the underlying type, it will be used to encode each element of the array. The second parameter is the size of the array: a Just-number for a fixed-size array (e.g. address[32]) or Nothing for a variable-length array (e.g. string[]).

The encoding in both cases is the keccak256 of the concatenation of the content's encodings.

If the array is fixed-sized, then the actual value's length will be validated upon encoding.

Struct handle

structValueHandle

Arguments

:: forall s structName. SName structName
The name of the struct. Use -XTypeApplications to declare constants here, e.g. SName @"Person".

-> [SignableStructMemberHandle s]
The list of the struct's member handles.

-> SignableValueHandle s

This handle encodes a struct type.

The encoding of a struct is keccak256 of the concatenation of the struct's typehash and each of its value's encoding in their definition order.

An example usage of this function:

       data MyPerson = MyPerson
           { pname :: String
           , pwallet :: BS.ByteString
           }
           deriving (Show, Eq, Generic)

       myPersonSignableValueHandle :: SignableValueHandle MyPerson
       myPersonSignableValueHandle =
           structValueHandle
               (SName @"Person")
               [   SignableStructMemberHandle "name" pname (signableValueHandle @"string")
               ,   SignableStructMemberHandle "wallet" pwallet (signableValueHandle @"address")
               ]
   

data SignableStructMemberHandle s

A handle for a specific field of a struct, to be used in structValueHandle.

Constructors

forall a. SignableStructMemberHandle

Fields

Text

The name of the field in Solidity.

(s -> a)

The accessor function (selector) of the field.

(SignableValueHandle a)

The encoding handle for the Haskell type of the field.

SignableValueHandle internals

data SignableValueHandle a

The handle that defines how a given Haskell value shall be encoded into a Solidity-compatible representation.

Normally, you should use one of the pre-defined handle factories, and possibly contramap or contramapSignableEither them onto your type.

EthType is a declaration of the Solidity type. Atomic types are declared completely, while struct types are declared only by a reference into a "type context". This allows multiple struct types to recursively refer to each other, a feature that's explicitly supported by EIP-712.

The encoding is done in two phases:

Firstly, we collect all struct definitions that are used in the root value to create the TypeContext.

Then, we actually traverse the given object, while resolving the struct-referencing EthTypes with the constructed TypeContext.

When building the type context, recursive struct definitions can lead to a value recursion in Haskell. For this reason, the primary source for a single handle is a value of TypeContextBuilder, which is a single struct definition followed by a list of its dependencies. When traversing the TypeContextBuilders, we can detect this kind of recursion by an explicit check and thus avoid getting stuck in an infinite loop when constructing the singular TypeContext object.

The Solidity value representation uses quite a bit of type magic to ensure consistency, but it's still not enough to prove that a struct within a TypeContext will exactly match the TypeContextBuilder that it came from. For this reason, signableValueEmbed takes a separate EthType, which may be different from the one given in signableValueEthType, and is allowed to report a mismatch between the Haskell type and the Solidity representation though the MonadFail instance of the Embed monad.

Constructors

SignableValueHandle

Fields

signableValueEthType :: Some EthType

The Solidity type that the values of this Haskell type should be mapped to.

signableValueTypeContextBuilder :: Maybe TypeContextBuilder

A "builder" for the type context. If this handle corresponds to a struct, this will be Just the definition of this struct and a list of its members' dependencies. Otherwise, this will be Nothing.

signableValueTypeContext :: Some TypeContext

The cached result of calling executeTypeContextBuilder on this handle's builder.

signableValueEmbed :: forall ctx rep. EthType rep -> a -> Embed ctx (EthValue ctx rep)

The encoder function. It's given a TypeContext as part of the Embed monad, an EthType and a Haskell value, and should either encode the given value according to the given EthType, or report a failure through the monad.

contramap :: Contravariant f => (a' -> a) -> f a -> f a'

contramapSignableEither :: (a -> Either String b) -> SignableValueHandle b -> SignableValueHandle a

A generalization of contramap @SignableValueHandle that allows you to perform run-time checks and fail gracefully by returning a Left. An error produced in this way will be annotated with a JSONPath-like location in the object where it happened.

If the conversion is total, you can use contramap instead.

data TypeContextBuilder

A builder of a TypeContext. This structure is specifically designed to allow value recursion, so that a definition like this would be legal:

       data FoobarTree = FoobarTree
           { stData :: Foobar
           , stChildren :: [FoobarTree]
           }

       foobarTreeTCB :: TypeContextBuilder
       foobarTreeTCB =
           TypeContextBuilder
               (Some $ SName @"FoobarTree")
               (Some $
                   MemberDef
                       (EthTypeStruct (SName @"Foobar"))
                       "data"
                 :/
                   MemberDef
                       (EthTypeArray (EthTypeStruct (SName @"FoobarTree")) Nothing)
                       "children"
                 :/
                   HEnd
               )
               [ foobarTCB
               , foobarTreeTCB -- value recursion
               ]
   

Constructors

TypeContextBuilder

Fields

(Some SName)

This struct's Solidity name.

(Some StructDef)

This struct's Solidity definition.

[TypeContextBuilder]

This struct's members' dependencies.

executeTypeContextBuilder :: TypeContextBuilder -> Some TypeContext

Fold a TypeContextBuilder into a TypeContext.

Value embedding monad

newtype Embed ctx a

A monad that is used for the main phase of the value encoding. Its structure can be represented by this stack:

       Embed ctx

     is-semantically-equivalent-to

       ReaderT (TypeContext ctx) -- The type context that all of the struct-referring EthTypes point to
     .
       ReaderT String -- The current position in the value, for error messages
     .
       ExceptT String -- Type mismatch errors
   

Constructors

Embed

Fields

withEmbed :: forall r. TypeContext ctx -> ShowS -> (String -> r) -> (a -> r) -> r

runEmbed :: TypeContext ctx -> Embed ctx a -> Either String a

The driver of the Embed actions.

withEmbedLocation :: String -> Embed ctx a -> Embed ctx a

local-like function that appends the given string to the current location.

withEmbedLocationArrayElement :: Int -> Embed ctx a -> Embed ctx a

withEmbedLocation specialization that appends an array index.

withEmbedLocationStructMember :: Text -> Embed ctx a -> Embed ctx a

withEmbedLocation specialization that appends a member access.

askTypeContext :: Embed ctx (TypeContext ctx)

Retrieve the current type context.

getStructDef :: SName typename -> Embed ctx (StructDef (Lookup typename ctx))

Lookup a struct definition by its name in the current type context.

Generic implementation internals

class GSignableRep (mods :: [SigningModifier]) s g where

Helper typeclass for the generic implementation of SignableValue.

The role of GSignableRep is to collect information about the fields of the structure. This list of fields is then used as a parameter to structValueHandle to make the complete handle.

Type parameter s stands for the original Haskell type, which the handle is being built for; and g is a GHC.Generics-representation of a part within the original value.

Methods

gSignableRepMemberHandles :: (s -> g ()) -> [SignableStructMemberHandle s]

type family ModifiedFieldName hsel mods where ...

ModifiedFieldName hsel mods is the Solidity field name that mods assign to selector hsel.

Equations

ModifiedFieldName hsel ('SigningFieldModifier hsel mbName mbType ': otherMods) = SafeUnify mbName (ModifiedFieldName hsel otherMods)

ModifiedFieldName hsel (_ ': otherMods) = ModifiedFieldName hsel otherMods

ModifiedFieldName hsel '[] = 'Nothing

type family ModifiedFieldType hsel mods where ...

ModifiedFieldType hsel mods is the wrapper type that mods assign to selector hsel.

Equations

ModifiedFieldType hsel ('SigningFieldModifier hsel mbName mbType ': otherMods) = SafeUnify mbType (ModifiedFieldType hsel otherMods)

ModifiedFieldType hsel (_ ': otherMods) = ModifiedFieldType hsel otherMods

ModifiedFieldType hsel '[] = 'Nothing

type family ValidateModifiers mods rep where ...

ValidateModifiers mods rep checks that the modifier list mods:

is internally consistent: it does not assign different names or different types to the same field, and

is applicable to the generic representation rep: it does not assign to fields that are not actually present in the structure.

If the checks pass, it reduces to 'Nothing, otherwise it's 'Just with the description of the failure as a type-level ErrorMessage.

Equations

ValidateModifiers (mod ': rest) rep = (ValidateModifierApplicable mod rep <|> ValidateModifierCompatible mod rest) <|> ValidateModifiers rest rep

ValidateModifiers '[] rep = 'Nothing

type family AssertValidation mbx r where ...

AssertValidation mbx r reduces to r if mbx ~ 'Nothing, otherwise it's TypeError.

It's intended to be used with ValidateModifiers type family, in the form AssertValidation (ValidateModifiers mods rep) r.

The reason it's split into a separate type family is to allow a form of validation that does not throw a type error on failure: ValidateModifiers mods rep ~ 'Nothing. This is sometimes used in conjunction with the "throwing" version to avoid duplicate error messages.

Equations

AssertValidation 'Nothing r = r

AssertValidation ('Just msg) r = TypeError msg