Data.SignTyped.Crypto

Description

We use Crypto.Secp256k1.Recovery as a back-end for the cryptographic operations.

Documentation

type MessageHash = Msg

type Signature = ByteString

type Address = ByteString

pubKeyToAddress :: Ctx -> PubKey -> Address

Derive an Ethereum-compatible address from a public key.

messageHashSign

Arguments

:: Ctx

-> SecKey

-> Maybe Int
chainId

-> MessageHash

-> Signature

Sign a given message with a given secret key. The signature produced is the 65-byte (v,r,s) recoverable signature.

If a chainId is provided, it will be encoded to the v byte of the signature according to EIP-155, as an offset of (chainId * 2 + 36). Otherwise, the default offset of 27 is used.

messageHashRecover

Arguments

:: Ctx

-> Maybe Int
chainId

-> MessageHash

-> Signature

-> Maybe Address

Recover an address from a given message and a given (v,r,s) signature.

For the v component, the default offset of 27 is always allowed. If a chainId is provided, its offset of (chainId * 2 + 36) will also be tried.

In the special case of chainId = 124 + 128 k, the value of v = 28 becomes ambiguous — it could mean both the odd parity with the default offset 27, and the even parity with the chainId-specific offset of 28. In this case, this function interprets it as the latter (the chainId-specific offset takes priority).

messageHashVerify :: Ctx -> Maybe Int -> MessageHash -> Signature -> Address -> Bool

Verify that a given (v,r,s) signature on a given message corresponds to a given address.

This function simply recovers the address with messageHashRecover and compares it to the one given.