Data.SignTyped.EIP712Domain

Documentation

type EIP712DomainSymbol = "EIP712Domain"

To avoid typos, use this constant for the canonical name of the domain struct, instead of typing the literal directly.

data EIP712Domain (opt :: EIP712DomainStructure)

The domain structure, from which the domain separator is derived.

The parameter configures the presence and Haskell representation of its fields.

The equality constraint is there to guide the type inference.

Constructors

opt ~ 'EIP712DomainStructure (EIP712DomainStructureNameType opt) (EIP712DomainStructureVersionType opt) (EIP712DomainStructureChainIdType opt) (EIP712DomainStructureVerifyingContractType opt) (EIP712DomainStructureSaltType opt) => EIP712Domain

Fields

eip712DomainName :: FMaybe (EIP712DomainStructureNameType opt)

eip712DomainVersion :: FMaybe (EIP712DomainStructureVersionType opt)

eip712DomainChainId :: FMaybe (EIP712DomainStructureChainIdType opt)

eip712DomainVerifyingContract :: FMaybe (EIP712DomainStructureVerifyingContractType opt)

eip712DomainSalt :: FMaybe (EIP712DomainStructureSaltType opt)

data SomeEIP712Domain

A wrapper that combines a EIP712Domain with its KnownEIP712DomainStructure instance.

Constructors

forall opt.KnownEIP712DomainStructure opt => SomeEIP712Domain (EIP712Domain opt)

data EIP712DomainStructure

In the EIP-712 spec, the fields of the EIP712Domain struct are optional. If some fields are not necessary in a specific case, they can be removed from the type definition entirely. Since we set up the SignableValue machinery to have struct definitions depend only on Haskell types, not values, that means each possible configuration of the domain fields must be a separate Haskell type. This structure is the type-level parameter that distinguishes between these configurations.

In addition, we allow the Haskell representation of each field to be any Haskell type that has a SignableValue instance for the appropriate Solidity typename.

Constructors

EIP712DomainStructure

Fields

eip712DomainStructureNameType :: Maybe Type

eip712DomainStructureVersionType :: Maybe Type

eip712DomainStructureChainIdType :: Maybe Type

eip712DomainStructureVerifyingContractType :: Maybe Type

eip712DomainStructureSaltType :: Maybe Type

type family EIP712DomainStructureNameType opt where ...

Equations

EIP712DomainStructureNameType ('EIP712DomainStructure b _ _ _ _) = b

type family EIP712DomainStructureVersionType opt where ...

Equations

EIP712DomainStructureVersionType ('EIP712DomainStructure _ b _ _ _) = b

type family EIP712DomainStructureChainIdType opt where ...

Equations

EIP712DomainStructureChainIdType ('EIP712DomainStructure _ _ b _ _) = b

type family EIP712DomainStructureVerifyingContractType opt where ...

Equations

EIP712DomainStructureVerifyingContractType ('EIP712DomainStructure _ _ _ b _) = b

type family EIP712DomainStructureSaltType opt where ...

Equations

EIP712DomainStructureSaltType ('EIP712DomainStructure _ _ _ _ b) = b

class (KnownCMaybe (SignableValue "string") (EIP712DomainStructureNameType opt), KnownCMaybe (SignableValue "string") (EIP712DomainStructureVersionType opt), KnownCMaybe (SignableValue "uint256") (EIP712DomainStructureChainIdType opt), KnownCMaybe (SignableValue "address") (EIP712DomainStructureVerifyingContractType opt), KnownCMaybe (SignableValue "bytes32") (EIP712DomainStructureSaltType opt), KnownCMaybe Show (EIP712DomainStructureNameType opt), KnownCMaybe Show (EIP712DomainStructureVersionType opt), KnownCMaybe Show (EIP712DomainStructureChainIdType opt), KnownCMaybe Show (EIP712DomainStructureVerifyingContractType opt), KnownCMaybe Show (EIP712DomainStructureSaltType opt)) => KnownEIP712DomainStructure opt

This typeclass is a "reified type synonym" for its context. Unlike a normal type synonym, however, this name can also be used partially applied, for example, as a parameter to KnownCMaybe.

eip712DomainEthType :: EthType ('EthReprStruct EIP712DomainSymbol)

The canonical EthType for EIP712Domain.

data family FMaybe (m :: Maybe Type) :: Type

The type family for a conditionally present value.