Data.SignTyped.Hashing

Documentation

hashEthValue :: TypeContext ctx -> EthType rep -> EthValue ctx rep -> ByteString

Produce the hash of the data to be signed.

This corresponds to the encodeData operation of the specification.

hashEthStruct :: TypeContext ctx -> SName typename -> HList (EthValue ctx) (Lookup typename ctx) -> ByteString

Produce the hash of the struct value to be signed.

This corresponds to the hashStruct operation of the specification.

structLocalContext :: TypeContext ctx -> SName typename -> Map Text ByteString

Collects all structs transitively used in defining the current one, including itself, and returns a map from struct names to their signatures.

For example, given this set of definitions in the type context:

       struct Person {
           string name;
           address wallet;
       }

       struct Mail {
           Person from;
           Person to;
           string contents;
       }
   

This function will return these results:

       structLocalContext tc (SName @"Person") =
           fromList
               [("Person", "Person(string name,address wallet)")]

       structLocalContext tc (SName @"Mail") =
           fromList
               [("Mail",   "Mail(Person from,Person to,string contents)"),
                ("Person", "Person(string name,address wallet)")]
   

This function is used as part of structTypeFullSignature.

structTypeFullSignature :: TypeContext ctx -> SName typename -> ByteString

Returns the argument of the type hash of the selected struct.

This corresponds to the encodeType operation in the specification.