Data.SignTyped.Json

Description

This modules defines the JSON format of Solidity types and objects.

For types, we define an intermediate representation, composed of conventional datas with standard Aeson instances, and functions for conversion between this intermediate representation and Data.SignTypes.Structure's objects.

For values, we convert between EthValue's and JSON directly.

Unmarked type context representation

type UnmarkedTypeContext = Map Text UnmarkedStructDef

A representation of a type context without the type-level markings, used as an intermediate between the marked TypeContext and completely untyped Value.

type UnmarkedStructDef = [UnmarkedMemberDef]

data UnmarkedMemberDef

Constructors

UnmarkedMemberDef

Fields

_type :: UnmarkedEthType

_name :: Text

type UnmarkedEthType = Text

Serializing into the unmarked representation

toUnmarkedTypeContext :: TypeContext ctx -> UnmarkedTypeContext

toUnmarkedStructDef :: StructDef reps -> UnmarkedStructDef

toUnmarkedMemberDef :: MemberDef rep -> UnmarkedMemberDef

toUnmarkedEthType :: EthType rep -> UnmarkedEthType

Parsing from the unmarked representation

parseUnmarkedTypeContext :: UnmarkedTypeContext -> Parser (Some TypeContext)

Convert an unmarked type context into a proper one.

This function also verifies that all types that are used in the context are either defined in the same context, or are built-in Solidity types.

parseUnmarkedStructDef :: UnmarkedStructDef -> Parser (Set Text, Some StructDef)

Parse an unmarked struct definition.

This function also verifies that the there are no members with the same name.

In addition to the proper struct definition, this function also returns the set of custom types that are used in this struct.

parseUnmarkedMemberDef :: UnmarkedMemberDef -> Parser (Set Text, Some MemberDef)

Parse an unmarked struct member definition.

In addition to the proper member definition, this function also returns the set of custom types that are used in the member.

parseUnmarkedEthType :: Text -> Parser (Set Text, Some EthType)

Parse an unmarked type definition.

In addition to the proper type definition, this function also returns the set of custom types that are used here.

Serializing and parsing Solidity values

ethValueToJSON :: TypeContext ctx -> EthType rep -> EthValue ctx rep -> Value

Implementation of toJSON for a typed Solidity object.

ethValueToEncoding :: TypeContext ctx -> EthType rep -> EthValue ctx rep -> Encoding

Implementation of toEncoding for a typed Solidity object.

ethValueParseJSON :: TypeContext ctx -> EthType rep -> Value -> Parser (EthValue ctx rep)

Implementation of parseJSON for a typed Solidity object.

null is accepted, and corresponds to defaultEthValue.

Auxiliary functions

isJsonRepresentableInteger :: Integer -> Bool

IEEE double-presicion numbers can represent integers within the range [-2^52 .. 2^52] exactly.

This function tests whether the given integer belongs to this range.