Data.SignTyped.Message

Documentation

data Message

A data structure that combines the message to sign, its domain and their type information.

It corresponds directly to the TypedData parameter of the eth_signTypedData request, and its JSON instances map directly to that.

As an emergent generalization, this structure allows the signable object to be any Solidity type with a defined encodeData, not just structs. This includes atomic types (such as uint256 and bytes) and arrays (e.g. bytes[4][]).

Constructors

forall ctx primRep. Message

Fields

messageTypeContext :: TypeContext ctx

messageDomainValue :: EthValue ctx ('EthReprStruct EIP712DomainSymbol)

messagePrimaryType :: EthType primRep

messagePrimaryValue :: EthValue ctx primRep

makeMessage :: forall a ethType. SignableValueDefault ethType a => SomeEIP712Domain -> a -> Either String Message

Given a domain and a signable object, encode them into a Message.

This function uses a SignableValueDefault instance to define the exact encoding. To choose a different encoding, use makeMessageWith.

makeMessageWith :: forall a. SignableValueHandle a -> SomeEIP712Domain -> a -> Either String Message

Given a domain and a signable object, encode them into a Message.

This function uses an explicitly given SignableValueHandle to define the exact encoding. To use the default encoding for the given type, you can use makeMessage instead.

messageChainId :: Message -> Maybe Int

Retrieve the chainId from the message's domain, if present.

hashMessage :: Message -> MessageHash

Calculates the final hash of the message, as given by the expression:

       keccak256("\x19\x01" ‖ domainSeparator ‖ hashStruct(message))
   

signMessage :: Ctx -> SecKey -> Message -> Signature

Sign a message with the given secret key.

recoverMessageSigner :: Ctx -> Message -> Signature -> Maybe Address

Recovers the signer's address from a message and a signature.