Data.SignTyped.Structure

Description

This modules defines a runtime representation of strictly typed Solidity values.

This representation is composed of three parts:

A TypeContext gives a mapping from type names to their declarations for custom types (structs).

An EthType gives a type definition. Custom types here are only referred by name, instead of being contained in their entirety, so the exact representation of a value will depend on both a TypeContext and an EthValue.

An EthValue is a type family that holds a Solidity object of a given Solidity type.

A mediator between the EthType and EthValue is EthRepr, which is a stripped type-level version of EthType. It contains only the information relevant to choosing the appropriate Haskell type for representation within an EthValue. And just like with EthType, custom types are only referred by their names in EthRepr too, and similarly require an addition of a type context (on a type level) to resolve fully.

Type definitions

data EthType rep where

A GADT defining a Solidity type. The type marker is an EthRepr selecting the appropriate Haskell representation for the given Solidity type.

Constructors

EthTypeAtomic
Atomic types are defined fully in-place.

They each have specific encodings and JSON mappings.

Fields

:: AtomicEthType t

An in-place definition for the type.

-> EthType ('EthReprAtomic t)

EthTypeArray
Array types are composed of a base type and an optional size.

The encoding is keccak256 of the concatenation of the elements' encodings.

In JSON, an array is represented as a JSON array.

Fields

:: EthType rep

Base (element) type.

-> Maybe Int

Array size. If present, the value must have this exact number of elements. Otherwise, any number of elements is permitted.

-> EthType ('EthReprArray rep)

EthTypeStruct
Structs are kept as references, to be fully resolved later with a TypeContext.

The encoding is keccak256 of a concatenation of the struct's type hash (dependent on its definition in the type context) and the encodings of its member values.

In JSON, a struct is represented by a JSON object, with keys corresponding to the struct members' names.

Fields

:: SName typename

The name of the struct, as a singleton.

-> EthType ('EthReprStruct typename)

data AtomicEthType t where

A GADT that gives specific definitions for atomic types.

Constructors

AtomicEthTypeBool :: AtomicEthType Bool
Represented by Haskell Bool: Solidity bool.

A bool is encoded as a uint256, with False and True corresponding respectively to 0 and 1.

In JSON, the literals false and true are used.

AtomicEthTypeInteger :: AtomicIntegerSignedness -> AtomicIntegerBitSize -> AtomicEthType Integer
Represented by Haskell Integer: Solidity intN and uintN.

When a number is encoded, it is sign-extended to 256 bits (32 bytes) and written out in big-endian order.

When serialized into JSON, the representation is dependent on the value's magnitude: if it is small enough to fit in the exact-integer-range of IEEE Double (which JSON numbers are usually implemented as), then the number is written as-is. Otherwise, it is converted to hexadecimal and written as a JSON string with preceding 0x.

On input from JSON, either form for any magnitude is allowed. Since Haskell's aeson always parses numbers precisely, it does not suffer from Double's limitations.

AtomicEthTypeBytes :: AtomicBytesSubtype -> AtomicEthType ByteString
Represented by Haskell ByteString: Solidity address, bytesN and bytes.

The single parameter further distinguishes between these variants, controls the way values are encoded for signing and mapped to JSON for sign requests, and can also impose restrictions on the value.

AtomicEthTypeString :: AtomicEthType Text
Represented by Haskell Text: Solidity string.

The encoding is keccak256 of the UTF-8 encoding of the string.

When serialized into JSON, the string is written out in literal text, unless it starts with 0x or 0X or contains an embedded zero, in which case it's encoded in UTF-8 and written in hexadecimal, like bytes.

data AtomicIntegerSignedness

Chooses between uintN and intN types.

Constructors

AtomicIntegerUnsigned

AtomicIntegerSigned

type AtomicIntegerBitSize = Int

data AtomicBytesSubtype

A subcategory of a bytes-represented type.

Constructors

AtomicBytesSubtypeAddress
Solidity address.

The bytestring's length must be exactly 20.

Upon encoding, the bytes are left-padded until the length of 32.

In JSON, the addresses are represented as hexadecimal strings with the mixed-case checksumming (ERC-55). The casing is strictly verified on input.

AtomicBytesSubtypeShortBytes Int
Solidity bytesN, with the N given in the single parameter.

The bytestring's length must be exactly N.

Upon encoding, the bytes are right-padded until the length of 32.

In JSON, the bytes are written out as a hexadecimal string. The input is case-insensitive. On output, lowercase digits are produced.

AtomicBytesSubtypeLongBytes
Solidity bytes.

There are no restrictions on the bytestring.

The encoding is keccak256 of the bytes.

In JSON, the bytes are written out as a hexadecimal string. The input is case-insensitive. On output, lowercase digits are produced.

ethTypeName :: EthType rep -> Text

Converts the marked type definition into a simple string.

For the inverse conversion, see parseUnmarkedEthType.

data EthRepr

A kind for selecting a Haskell representation for a Solidity object.

Constructors

EthReprAtomic Type

EthReprArray EthRepr

EthReprStruct Symbol

Typed values

data family EthValue (ctx :: TMap Symbol [EthRepr]) (rep :: EthRepr)

Haskell representation of a Solidity object of the given type.

The first parameter matches the representation with a TypeContext, and the second with an EthType.

To inspect an arbitrary EthValue, you will need its corresponding TypeContext and EthType. As examples of processing arbitrary objects using this type information, you can look at the sources of defaultEthValue, ethValueToEncoding, ethValueParseJSON, hashEthValue and other functions from this package.

defaultEthValue :: TypeContext ctx -> EthType rep -> EthValue ctx rep

Generate a default value for a given type:

bool defaults to False.

Integer types default to 0.

Fixed-length strings (address and bytesN) are filled with zero bytes.

Variable-length strings (bytes and string) are defaulted to the empty string.

Fixed-length arrays are filled with the copies of the base type's default value.

Variable-length arrays are defaulted empty.

Structs have each of their member defaulted.

eqEthValue :: TypeContext ctx -> EthType rep -> EthValue ctx rep -> EthValue ctx rep -> Bool

Checks equality of two Solidity objects of the same type.

Since inspecting EthValues requires type information to be passed along with the value, we cannot do this in the standard (==) operator.

compareEthValue :: TypeContext ctx -> EthType rep -> EthValue ctx rep -> EthValue ctx rep -> Ordering

Compares the ordering of two Solidity objects of the same type.

Since inspecting EthValues requires type information to be passed along with the value, we cannot do this in the standard compare operator.

Struct definitions

type StructDef = HList MemberDef

data MemberDef rep where

Structs are defines by their members, and each member is defined by its name and type.

The member type's representation marker is exposed as the marker of the member itself. When collected into an HList, its combined marker gives the representation of the entire structure, as a list of representations of each of its members.

Constructors

MemberDef :: EthType rep -> Text -> MemberDef rep

Type context

newtype TypeContext ctx

A type context is an HMap from struct names to their definitions (member lists).

In the context of Haskell's type system, a TypeContext is a singleton — for each instance, the marker ctx will take on a unique type of kind TMap SName StructDef. Combined with EthType's markers, this makes this representation of Solidity objects in Haskell almost an instance of dependent typing — where the values of the TypeContext and EthType determine the exact type of an EthValue.

With the marker "hidden" under an existential, Some TypeContext practically becomes a normal map. To that end, it posesses an instance of Monoid (Some TypeContext), which allows one to construct a type context from individual definitions.

Constructors

TypeContext

Fields

typeContextMap :: HMap SName StructDef ctx

emptyTypeContext :: Some TypeContext

Produce an empty type context.

singletonTypeContext :: SName typename -> StructDef reps -> Some TypeContext

Produce a type context with a single definition.

lookupStructDef :: TypeContext ctx -> SName typename -> Maybe (StructDef (Lookup typename ctx))

Perform a safe lookup in a type context. If the given key is absent, Nothing is returned.

withStructDef :: TypeContext ctx -> SName typename -> (StructDef (Lookup typename ctx) -> r) -> r

Perform a strict lookup in a type context. If the given key is absent, an error is raised.

The continuation-passing style of this function allows one to put an arbitrary expression dependent on the check, so that even if the struct definition itself ends up being unused due to lazyness, the key-presence check will be performed anyway.

Printing for debugging

showsEthValue :: TypeContext ctx -> EthType rep -> EthValue ctx rep -> ShowS

Prints a Solidity object into a ShowS-string. This is intended solely for debugging purposes, so the output is not guaranteed to conform to any format or even be unambiguous.

If you want to properly serialize a Solidity object, you should use ethValueToJSON or ethValueToEncoding instead.

showsEthStructDef :: SName typename -> StructDef reps -> ShowS

Prints a struct definition into a ShowS-string. This is intended solely for debugging purposes, so the output is not guaranteed to conform to any format or even be unambiguous.

If you want to properly serialize a struct definition, you should instead use toUnmarkedStructDef, followed by any of the Aeson's serialization routines.

showsTypeContext :: TypeContext ctx -> ShowS

Prints a type context into a ShowS-string. This is intended solely for debugging purposes, so the output is not guaranteed to conform to any format or even be unambiguous.

If you want to properly serialize a type context, you should instead use toUnmarkedTypeContext, followed by any of the Aeson's serialization routines.

Unmarked Solidity objects for testing

data SomeEthValue

A Solidity object wrapped together with its type.

This wrapper is intended to be used in test suites, where assertions expect the objects to have standard Show and Eq instances, for example as arguments to shouldBe.

Constructors

forall ctx repr. SomeEthValue (TypeContext ctx) (EthType repr) (EthValue ctx repr)

Validating type invariants

validateEthValue :: TypeContext ctx -> EthType rep -> EthValue ctx rep -> [String]

Performs validation checks on the given object and returns the list of errors.

This function checks for additional conditions imposed by the Solidity type definition that are not part of the representation's Haskell type itself, such as the integer values being in range of their types, or fixed-size arrays having a proper number of elements.

The default handles from SignableValue, as well as ethValueParseJSON perform their own validation to produce only correct Solidity objects, so this specific function is better suited as a debug tool rather than a routine checker in the main code.

validateEthValueWorker :: ShowS -> TypeContext ctx -> EthType rep -> EthValue ctx rep -> [String] -> [String]

Internal recursive function for validateEthValue.

The first argument is a ShowS-encoded prefix of the value's location (e.g. showString ".elems[3].first").

The partial list is returned as a difference-list.

The default entrypoint is defined like this:

       validateEthValue tc ethType ethValue =
           validateEthValueWorker (showString "$") tc ethType ethValue []
   

isValidInteger :: AtomicIntegerSignedness -> AtomicIntegerBitSize -> Integer -> Bool

Check that the given integer fits within the given type.