Data.SignTyped.Type.HList

Documentation

data HList f ts where

A heterogeneous list.

The first parameter is a unary type constructor, which is applied to each of the elements of the second parameter to obtain the full type of each of the HList's elements.

For example, a value of type HList f '[t1, t2, t3] contains three objects of types (f t1, f t2, f t3).

Constructors

(:/) :: f t -> HList f rest -> HList f (t ': rest) infixr 4

HEnd :: HList f '[]

hmap :: (forall t. f t -> g t) -> HList f ts -> HList g ts

Heterogeneous map. The mapping function is a natural transformation.

For example:

       hmap
           (maybe (Left "no value") Right
               :: forall t. Maybe t -> Either String t)
           (Just 10 :/ Just "text"  :/ Nothing @Float  :/ HEnd
               :: HList Maybe '[Int, String, Float])
     ==
           Right 10 :/ Right "text" :/ Left "no value" :/ HEnd
               :: HList (Either String) '[Int, String, Float]

   

htraverse :: Applicative m => (forall t. f t -> m (g t)) -> HList f ts -> m (HList g ts)

Heterogeneous traverse.

For example:

       htraverse
           (maybe ([Left "no value"]) (\x -> [Left "removed", Right x])
               :: forall t. Maybe t -> [Either String t])
           (Just 10         :/ Just "text"    :/ Nothing @Float  :/ HEnd
               :: HList Maybe '[Int, String, Float])
     ==
           [ Left "removed" :/ Left "removed" :/ Left "no value" :/ HEnd
           , Left "removed" :/ Right "text"   :/ Left "no value" :/ HEnd
           , Right 10       :/ Left "removed" :/ Left "no value" :/ HEnd
           , Right 10       :/ Right "text"   :/ Left "no value" :/ HEnd
           ]
               :: [HList (Either String) '[Int, String, Float]]

   

hfoldr :: (forall t. f t -> r -> r) -> r -> HList f ts -> r

Heterogeneous foldr.

For example:

       hfoldr
           (maybe ("n" ++) (\_ -> ("j" ++))
               :: forall t. Maybe t -> String -> String)
           ""
           (Just 10 :/ Just "text"  :/ Nothing @Float  :/ HEnd
               :: HList Maybe '[Int, String, Float])
     ==
           "jjn"

   

hfoldMap :: Monoid m => (forall t. f t -> m) -> HList f ts -> m

Heterogeneous foldMap.

hzip :: HList f ts -> HList g ts -> HList (Product f g) ts

Heterogeneous zip.

For example:

       hzip
           (Just 10         :/ Just "text"    :/ Nothing         :/ HEnd
               :: HList Maybe '[Int, String, Float])
           (Const "a"       :/ Const "b"      :/ Const "c"       :/ HEnd
               :: HList (Const String) '[Int, String, Float])
     ==
           Pair (Just 10) (Const "a") :/ Pair (Just "text") (Const "b") :/ Pair Nothing (Const "c") :/ HEnd
               :: HList (Product Maybe (Const String)) '[Int, String, Float]

   

hzipWith :: (forall t. f t -> g t -> h t) -> HList f ts -> HList g ts -> HList h ts

Heterogeneous zipWith.

For example:

       hzipWith
           ((\ma eb -> maybeToList ma <> rightToList eb)
               :: forall t. Maybe t -> Either () t -> [t])
           (Just 10         :/ Just "text"    :/ Nothing         :/ HEnd
               :: HList Maybe '[Int, String, Float])
           (Right 20        :/ Left ()        :/ Right 34.5      :/ HEnd
               :: HList (Either ()) '[Int, String, Float])
     ==
           [10, 20] :/ ["text"] :/ [34.5] :/ HEnd
               :: HList [] '[Int, String, Float]

   

htoList :: HList (Const a) ts -> [a]

Convert an HList of Const into a normal list.

hfromList :: [a] -> Some (HList (Const a))

Convert a normal list into an HList of Const.

hsequenceSome :: [Some f] -> Some (HList f)

Convert a normal list of Somes into an HList.

The name refers to the fact that, similar to sequenceA, the effect of this function can be seen as a transposition of containers.

data Product (f :: k -> Type) (g :: k -> Type) (a :: k)

Lifted product of functors.

Constructors

Pair (f a) (g a)