Data.SignTyped.Type.HMap

Description

This module defines a singleton heterogeneous map.

This map provides a mapping that exists on both type-level and value-level. The keys of the map are singletons, where each distinct value has its own type marker. The values are arbitrary GADTs that use the type parameter as a marker. For each mapping from a key object and a value object, there is a corresponding mapping from the key's to the value's type markers inside the map's state token.

The state token itself is not constructive, it has no direct representation in the Haskell's type system. Instead, it is being kept as an abstract existential, which is consistently the same for the same object, but distinct for any two non-equivalent objects. The presence of a particular mapping inside the state token is asserted axiomatically by the HMap-operating functions.

Whenever a new HMap object is created, it is wrapped in a Some. This effectively makes GHC spawn a new existential type for each such creation.

Documentation

data TMap kk kv

A kind for state tokens of HMaps.

type family Lookup k ms where ...

A type family that represents a lookup result within a specific state token.

There is no constructive definition for this type family, its instances are produced axiomatically by HMap's functions.

data HMap kf vf s

A singleton heterogeneous map.

The last parameter is a type-level token corresponding to the map's value.

emptyHMap :: Some (HMap kf vf)

singletonHMap :: SingletonConstructor kf => kf kt -> vf vt -> Some (HMap kf vf)

unionHMapWithKey :: SingletonConstructor kf => (Some kf -> Some vf -> Some vf -> Some vf) -> Some (HMap kf vf) -> Some (HMap kf vf) -> Some (HMap kf vf)

lookupHMap :: forall kf kt vf s. SingletonConstructor kf => kf kt -> HMap kf vf s -> Maybe (vf (Lookup kt s))

foldMapHMapWithKey :: Monoid m => (forall kt. kf kt -> vf (Lookup kt s) -> m) -> HMap kf vf s -> m

foldrHMapWithKey :: (forall kt. kf kt -> vf (Lookup kt s) -> r -> r) -> r -> HMap kf vf s -> r

data Some f where

A wrapper for existential quantification.

Semantically, it's definition is:

       data Some (f :: k -> Type) = forall (a :: k). Some (f a)
   

It's primarily intended to be used with marker GADTs and singleton types, such as SName.

Bundled Patterns

pattern Some :: forall f. forall t. f t -> Some f