Data.SignTyped.Type.SCMaybe

Description

This module defines conditionally present polymorphic values. Unlike Prelude's Maybe, the condition is given at the type level. If the value is present, it can be of any type, but it must support the given constraint.

In the sign-typed-data library, it's used for the fields of EIP712Domain: depending on the application, some fields may be absent, and those that are present might use an application-specific representation, provided that there is a way to encode it into a Solidity object.

The S letter in this module stands for "singleton", the same as in SNat, SChar etc.

The C letter stands for "constrained", and refers to the fact that a non-empty type must support the given constraint.

The F letter in FMaybe simply means "type family" (as opposed to a GADT or a normal data).

Documentation

data SCMaybe (c :: k -> Constraint) (m :: Maybe k) where

A singleton marker for a constrained type-level Maybe.

An explicit forall on the SCJust constructor allows the existential a to be pattern-matched with a type-application pattern, for example:

       mbNatVal :: SCMaybe KnownNat mbx -> Maybe Integer
       mbNatVal SCNothing = Nothing
       mbNatVal (SCJust @x) = Just (natVal @x undefined)
   

Constructors

SCNothing :: SCMaybe c 'Nothing

SCJust :: forall a c. c a => SCMaybe c ('Just a)

class KnownCMaybe c m where

A class to automatically produce SCMaybes .

Methods

cmaybeSing :: SCMaybe c m

withCMaybe :: forall c m r. KnownCMaybe c m => (m ~ 'Nothing => r) -> (forall a. (m ~ 'Just a, c a) => r) -> r

A destructor for a context-provided SCMaybe, analogous to maybe, either and bool.

data family FMaybe (m :: Maybe Type) :: Type

The type family for a conditionally present value.

fmaybe :: forall c m r. KnownCMaybe c m => (m ~ 'Nothing => r) -> (forall a. (m ~ 'Just a, c a) => a -> r) -> FMaybe m -> r

A destructor for an FMaybe, which uses a KnownCMaybe from the context to determine the actual state of the value (present or absent) and retrieve the required constraint in the present case.