Data.SignTyped.Type.SName

Description

Singleton type for names that use Text for representation, in the style of GHC.TypeLits.

Documentation

data SName (sym :: Symbol) where

Singleton for a specific name.

To create a known SName constant, use SName @"str" or nameSing @"str".

To create an SName from a run-time value, use withSomeSName.

To obtain the value inside, use fromSName.

Bundled Patterns

pattern SName :: forall sym. () => KnownName sym => SName sym
An explicitly bidirectional pattern synonym that converts between SName values and KnownName constraints.

class KnownName sym where

A class for known names. Any known Symbol can be converted to a name.

Methods

nameSing :: SName sym

The singleton for the name.

withSName :: forall sym r. SName sym -> (KnownName sym => r) -> r

Convert a singleton into a typeclass instance.

In other words, use an SName value as material evidence of a KnownName.

withSomeSName :: Text -> (forall sym. SName sym -> r) -> r

Convert an arbitrary text into an SName.

fromSName :: SName sym -> Text

Extract the text value from an SName.

toSomeSName :: Text -> Some SName

Construct an SName from an arbitrary text.

data Symbol

(Kind) This is the kind of type-level symbols. Declared here because class IP needs it