Data.SignTyped.Type.TestOrdering

Documentation

class TestEquality f => TestOrdering f where

Methods

testOrdering :: f a -> f b -> OrderingI a b

data OrderingI (a :: k) (b :: k) where

Ordering data type for type literals that provides proof of their ordering.

Since: base-4.16.0.0

Constructors

LTI :: forall {k} (a :: k) (b :: k). Compare a b ~ 'LT => OrderingI a b

EQI :: forall {k} (a :: k). Compare a a ~ 'EQ => OrderingI a a

GTI :: forall {k} (a :: k) (b :: k). Compare a b ~ 'GT => OrderingI a b

class TestOrdering f => SingletonConstructor f

This typeclass declares that a type constructor produces singletons. That is, for each possible type argument a, there is at most one possible object of type f a.

In other words, the instances of this class must satisfy the law:

testOrdering (fa :: f a) (fb :: f b) == EQI    ===>    a !~ b

Examples of good instances of this class are the singletons from GHC.TypeLits, and also TypeRep and SName.

There are no additional methods to implement in this class, it is only used as a marker.