Data.SignTyped.Util

Description

An assortment of utility functions that are used across other modules.

Documentation

symbolText :: forall (s :: Symbol). KnownSymbol s => Text

Take KnownSymbol value as a Text.

keccak256 :: ByteString -> ByteString

keccak256 hash of a bytestring, as a bytestring.

This is the main hashing function used in Ethereum.

integer256 :: Integer -> ByteString

Serialize an integer into a 32-byte big-endian 2's complement bytestring.

       integer256 0      = unhex "0x0000000000000000000000000000000000000000000000000000000000000000"
       integer256 1      = unhex "0x0000000000000000000000000000000000000000000000000000000000000001"
       integer256 100    = unhex "0x0000000000000000000000000000000000000000000000000000000000000064"
       integer256 256    = unhex "0x0000000000000000000000000000000000000000000000000000000000000100"
       integer256 (-1)   = unhex "0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff"
       integer256 (-256) = unhex "0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff00"
   

readNumber :: Text -> Maybe Integer

Try to parse the string as an integer.

For simplicity, this re-uses the Read instance of Integer, which is quite permissive. For example, all of these strings are accepted as valid representations of the number (-10):

       "-10"
       "-0xa"
       "-0o12"
       "   -   10   "
       "\n\r\t\v\f   -10   "
       "(((-10)))"
   

This function is used to parse intN and uintN values from JSON, so that other programs where native JSON integers are limited by a relatively small number of bits could still produce JSONs with large integers by encoding them in hexadecimal and stringifying.

readChecksumAddress :: Text -> Maybe ByteString

Try to parse the string as an address, while also verifying the casing of its letter digits to be the correct ERC-55 checksum.

This function is used to parse address values.

readSafeTextString :: Text -> Maybe Text

Try to parse the string as a potentially hex-encoded string. More specifically, if the input string starts with 0x or 0X, it is parsed as a hex-encoded bytestring and then parsed from UTF-8. Otherwise, it is simply interpreted as a literal string.

       readSafeTextString "" = Just ""
       readSafeTextString "Some text" = Just "Some text"
       readSafeTextString "0x" = Just ""
       readSafeTextString "0x6f746865722074657874" = Just "other text"
       readSafeTextString "0x3078" = Just "0x"
       readSafeTextString "0xf09fa494" = Just "129300"
       readSafeTextString "0xff" = Nothing -- due to being invalid utf-8
   

This function is used to parse string values.

readHexBytes :: Text -> Maybe ByteString

Try to parse the string as a hex-encoded bytestring.

This function is used to parse bytesN and bytes values.

encodeChecksumAddress :: ByteString -> Text

Encode an address in hexadecimal, while also applying the ERC-55 mixed-case checksumming.

This function is used to serialize address values.

encodeSafeTextString :: Text -> Text

Encode a text string in such a way that it can later be parsed unambiguously. If the input string starts with 0x or 0X, it will be hex-encoded, otherwise the bytes are UTF-8 decoded and simply output literally.

       encodeSafeTextString "" = ""
       encodeSafeTextString "Some text" = "Some text"
       encodeSafeTextString "0x is a prefix" = "0x3078206973206120707265666978"
   

This function is used to serialize string values.

encodeHexByteString :: ByteString -> Text

Encode a bytestring in hexadecimal.

This function is used to serialize bytesN and bytes values.

encodeHexInteger :: Integer -> Text

Encode an integer in hexadecimal.

This function is used to serialize huge intN and uintN values. This allows other programs where native JSON integers are limited by a relatively small number of bits to still parse and accept large integers without loss of precision.